+++ test/case-expandmultivalue.sh
@@ -1,0 +1,3 @@
+
+./xsvutils ./test/data/sample-3.tsv --header col1,col2 expandmultivalue --col col1 > ./var/test/expandmultivalue-1.tsv
+